sentence tokenization options
allow for external sentence tokenization. conceptually parallel to those implemented in the VADER plugin beginning here:
https://github.com/BUTTER-Tools/VADER/commit/c20d62e89de7df953da54df19adffce9c74c46a5

Note that this change allows users to how sentence tokenization happens (internal versus external). Part of the process of working through the methods for the "News media and shared representations: The development and validation of an automated approach to the detection of linguistic bias" project:
https://www.sshrc-crsh.gc.ca/results-resultats/recipients-recipiendaires/2020/insight_development_grants-subventions_developpement_savoir-eng.aspx